requirements.txt
pandas
sqlalchemy
pymysql
requests
streamlit
prophet
scikit-learn
apscheduler
python-dotenv
plotly
beautifulsoup4
